feat(console-settings): re-home Parent Domains as a granular #parent-domains section (drop the odd-one-out left-pane item)
Parent Domains was the lone child in SovereignSidebar's SETTINGS_SUB_NAV —
the only odd-one-out sub-left-pane menu item under Settings. Every other
settings surface is a granular, anchor-navigable section of the unified
/settings page (#organization, #sovereign, #dns, #domain-mode,
#marketplace, …). The founder wants Parent Domains to match that pattern
(reachable at /settings#parent-domains, like /settings#dns), exactly the
move Marketplace already got in Wave 5.

Changes (console settings UI only):
- Extract the inner content of ParentDomainsPage into a reusable,
  PortalShell-free ParentDomainsSection (same idiom as MarketplaceSection).
  Zero functional change: list pools, role/flip-status badges, inline
  propagation drawer, "+ Add another domain" modal, remove flow all move
  over verbatim, every data-testid preserved.
- SettingsPage: register { id: 'parent-domains' } in SECTIONS (drives the
  TOC + heading) and render <SectionCard id="parent-domains">
  <ParentDomainsSection /></SectionCard> right after Domain mode. Switch
  the SectionCard descriptions from fragile SECTIONS[N] index lookups to
  an id-based desc() helper so mid-array inserts can't drift.
- ParentDomainsPage becomes a thin PortalShell wrapper around
  ParentDomainsSection so the standalone /organizations/domains route
  (and the /parent-domains redirect that lands there) keep working
  byte-for-byte — no deep-link breaks.
- SovereignSidebar: drop the now-empty SETTINGS_SUB_NAV + the sub-nav
  expansion machinery (deriveActiveSettingsSubItem, settingsExpanded, the
  sub-list render). Settings is now a single flat link like every other
  top-level nav entry.
- SettingsPage.test: add 'parent-domains' (and the always-rendered
  'marketplace') to ALL_SECTIONS so the section-catalogue + TOC tests
  cover the new anchor.

Gates: tsc -b --noEmit clean; vite build clean. (SettingsPage.test +
ParentDomainsPage.test currently fail identically on pristine HEAD due to
a @tanstack/react-router 1.170.15 / test-harness incompatibility — a
pre-existing breakage on main, not introduced here.)

Refs #4089, Refs #825, Refs #3378, Refs #516
